// src/components/dialogs/network-modifications/two-windings-transformer/tap-changer-pane/ratio-tap-changer-pane/ratio-tap-changer.type.ts
/**
 * Copyright (c) 2025, RTE (http://www.rte-france.com)
 * This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/.
 */

export interface RatioTapChangerStepData {
    index: number;
    rho: number;
    r: number;
    x: number;
    g: number;
    b: number;
}

export interface RatioTapChangerFormInfos {
    lowTapPosition: number;
    tapPosition: number;
    highTapPosition: number;
    regulatingTerminalConnectableId: string;
    regulatingTerminalConnectableType: string;
    regulatingTerminalVlId: string;
    targetV: number;
    targetDeadband: number;
    isRegulating: boolean;
    hasLoadTapChangingCapabilities: boolean;
    steps: RatioTapChangerStepData[];
}